requirements.txt
requests
python-dotenv
pandas
sentence-transformers
faiss-cpu
streamlit
plotly